Se crea primeras pruebas de testing

+++ Back/JPA/src/main/java/com/ecoomerce/JPA/repositories/ColorRepository.java
@@ -3,6 +3,6 @@ package com.ecoomerce.JPA.repositories;
 import org.springframework.data.repository.CrudRepository;
 import com.ecoomerce.JPA.entitys.Color;
 
-public interface ColorRepository extends CrudRepository<Color, Long> {
+public interface ColorRepository extends CrudRepository<Color, Integer> {
 
 }
